requirements.txt
opencv-python
matplotlib
seaborn
pytorch-lightning>=1.4
debuggy
tqdm
